resolved dependency issue, can open the graphiQL interface, but the query always returns null

+++ inventory-service/src/main/java/com/urbanthreads/inventoryservice/model/Image.java
@@ -1,12 +1,11 @@
 package com.urbanthreads.inventoryservice.model;
-
-
 import jakarta.persistence.*;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.Set;
+
 
 @Setter
 @Getter
@@ -16,7 +15,7 @@ import java.util.Set;
 public class Image {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
